requirements.txt
asgiref==3.4.1
Django==3.2.7
django-colorfield==0.4.3
django-cors-headers==3.9.0
django-environ==0.7.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
djangorestframework-simplejwt==4.8.0
environ==1.0
gunicorn==20.1.0
inflection==0.5.1
PyJWT==2.1.0
pyserial==3.5
pytz==2021.1
sqlparse==0.4.2
